Added CityBuilder Widget.
It handles the update flow of the Sloboda events.

If you need to rerender your widget with new Sloboda state, then wrap it with CityBuilder.
If your widget changes the state of Sloboda and needs to render new state on its screen, then wrap it with CityBuilder and do not write any setState calls. Everything will be done automatically.